Keep 3.5 compat
