+++ tests/p36/conftest.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def pytest_ignore_collect():
+    if sys.version_info < (3, 6):
+        return True
